// apps/desktop/src/types/generated/ChangelogEntryDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChangelogChangeDto } from "./ChangelogChangeDto";

export type ChangelogEntryDto = {
  id: string;
  title: string;
  version: string | null;
  text: string | null;
  createdAt: number | null;
  changes: Array<ChangelogChangeDto>;
};